// src/generated/types/UpdateSoftwareDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type UpdateSoftwareDto = {
  /**
   * @type string | undefined
   */
  name?: string | undefined
  /**
   * @type string | undefined
   */
  requiredMemory?: string | undefined
  /**
   * @type string | undefined
   */
  requiredStorage?: string | undefined
}